functions/requirements.txt
# Core networking and utilities (keep these as they are or unpin if causing issues)
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
banks==2.1.3
beautifulsoup4==4.13.4
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
chromedriver-autoinstaller==0.6.4
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
cssselect==1.3.0
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
exceptiongroup==1.3.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.5.1
google-api-core==2.25.1
google-api-python-client==2.175.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.1.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
griffe==1.7.3
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
html2text==2024.2.26
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
lxml==6.0.0
lxml_html_clean==0.4.2
markdownify==1.1.0
MarkupSafe==3.0.2
marshmallow==3.26.1
msgpack==1.1.1
multidict==6.6.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.2.6
outcome==1.3.0.post0
oxylabs==2.0.0
packaging==25.0
pandas==2.2.3
pillow==11.3.0
platformdirs==4.3.8
playwright==1.53.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
PyJWT==2.10.1
pyparsing==3.2.3
pypdf==5.7.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-file==2.1.0
rsa==4.9.1
selenium==4.34.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
spider-client==0.0.27
SQLAlchemy==2.0.41
striprtf==0.0.26
tenacity==9.1.2
tiktoken==0.9.0
tinysegmenter==0.3
tldextract==5.3.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
websocket-client==1.8.0
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.1

# --- FIREBASE & CORE LLM/VECTOR DB PACKAGES (ENSURE THESE ARE PRESENT) ---
firebase-admin==6.9.0
firebase-functions

# Core LlamaIndex library - let pip choose a compatible version
llama-index

# Explicitly add the Pinecone vector store integration
llama-index-vector-stores-pinecone

# Explicitly add web reader (if needed for ingestion)
llama-index-readers-web

# Core OpenAI Python client
openai

# Core Pinecone Python client
pinecone-client

# --- REMOVE THESE SPECIFIC LLAMA-INDEX SUB-PACKAGES (as they should be covered by 'llama-index' main package or are not directly used in main.py) ---
# llama-cloud==0.1.30
# llama-cloud-services==0.6.41
# llama-index-agent-openai==0.4.12
# llama-index-cli==0.4.3
# llama-index-core==0.12.46         # REMOVE: llama-index should handle this
# llama-index-embeddings-openai==0.3.1
# llama-index-indices-managed-llama-cloud==0.7.8
# llama-index-instrumentation==0.2.0
# llama-index-llms-openai==0.4.7
# llama-index-multi-modal-llms-openai==0.5.1
# llama-index-program-openai==0.3.2
# llama-index-question-gen-openai==0.3.1
# llama-index-readers-file==0.4.9
# llama-index-readers-llama-parse==0.4.0
# llama-index-workflows==1.0.1
# llama-parse==0.6.41